# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tarn AI Evaluation API",
    "description": "AI suggests; the teacher decides. Every college sees only its own data.",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/booklets": {
      "post": {
        "tags": [
          "booklets"
        ],
        "summary": "Upload a booklet: one PDF, or the page images in order",
        "description": "`files` is one PDF, or JPEG/PNG images in page order (the order sent is the page order). The type is judged from the file's content. The booklet is queued for page cleaning (status `uploaded`). The same file hash within the college is refused with 409 unless `allow_duplicate` is true; a teacher may have a limited number of booklets waiting (429 beyond it).",
        "operationId": "upload_booklet_api_v1_booklets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_booklet_api_v1_booklets_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookletOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "description": "The same file was uploaded before (`duplicate_of` lists the booklets)"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "429": {
            "description": "You already have the maximum waiting",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "booklets"
        ],
        "summary": "Your college's booklets with their processing status, newest first",
        "operationId": "list_booklets_api_v1_booklets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mine",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only booklets I uploaded",
              "default": false,
              "title": "Mine"
            },
            "description": "Only booklets I uploaded"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "uploaded",
                    "processing",
                    "needs_retake",
                    "pages_ready",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookletPageOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}": {
      "get": {
        "tags": [
          "booklets"
        ],
        "summary": "One booklet with its pages, their measurements and retake reasons",
        "operationId": "get_booklet_api_v1_booklets__booklet_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookletDetailOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "booklets"
        ],
        "summary": "Delete a booklet: images, text and marks go; a content-free record stays",
        "operationId": "delete_booklet_api_v1_booklets__booklet_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/pages/{number}/image": {
      "get": {
        "tags": [
          "booklets"
        ],
        "summary": "The cleaned page image (or the original with `kind=original`)",
        "operationId": "page_image_api_v1_booklets__booklet_id__pages__number__image_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "cleaned",
                "original"
              ],
              "type": "string",
              "default": "cleaned",
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/jpeg": {},
              "image/png": {}
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/pages/{number}/text": {
      "get": {
        "tags": [
          "booklets"
        ],
        "summary": "What OCR read on one page: every line with every engine's reading and score",
        "description": "Lines keep every engine's reading, the selector's terms for each, the chosen one and whether the line is flagged for the teacher. Tables come as a `table` region followed by its cells (`parent_id`, `row`, `col`). Empty until the page has been read.",
        "operationId": "page_text_api_v1_booklets__booklet_id__pages__number__text_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageTextOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/pages/{number}/use-anyway": {
      "post": {
        "tags": [
          "booklets"
        ],
        "summary": "Go on with a page the quality gate flagged",
        "description": "Overrides the retake request for that page. When no flagged page is left the booklet moves to `pages_ready`. Recorded in the audit log.",
        "operationId": "use_page_anyway_api_v1_booklets__booklet_id__pages__number__use_anyway_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          },
          {
            "name": "number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookletOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_booklet_api_v1_booklets_post": {
        "properties": {
          "student_id": {
            "type": "string",
            "format": "uuid",
            "title": "Student Id"
          },
          "blueprint_id": {
            "type": "string",
            "format": "uuid",
            "title": "Blueprint Id"
          },
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files",
            "description": "One PDF, or images in page order"
          },
          "allow_duplicate": {
            "type": "boolean",
            "title": "Allow Duplicate",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "blueprint_id",
          "files"
        ],
        "title": "Body_upload_booklet_api_v1_booklets_post"
      },
...
      "BookletPageOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BookletOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "waiting": {
            "type": "integer",
            "title": "Waiting",
            "description": "Your booklets that are queued or being processed."
          },
          "max_waiting": {
            "type": "integer",
            "title": "Max Waiting",
            "description": "How many you may have waiting at once."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "waiting",
          "max_waiting"
        ],
        "title": "BookletPageOut"
      },
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageTextOut": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "text_read": {
            "type": "boolean",
            "title": "Text Read"
          },
          "needs_text": {
            "type": "boolean",
            "title": "Needs Text"
          },
          "ocr_failures": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ocr Failures"
          },
          "regions": {
            "items": {
              "$ref": "#/components/schemas/RegionOut"
            },
            "type": "array",
            "title": "Regions",
            "description": "In reading order; table cells after their table."
          }
        },
        "type": "object",
        "required": [
          "number",
          "text_read",
          "needs_text",
          "ocr_failures",
          "regions"
        ],
        "title": "PageTextOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /api/v1/auth/login",
        "scheme": "bearer"
      }
    }
  }
}